pyproject.toml
[project]
name = "llm-data-aug"
version = "0.1.0"
description = "Comparative study of Resampling and LLM for data augmentation for sentiment analysis tasks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "emoji-vietnamese>=0.6.6",
    "emoji>=2.14.1",
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pyvi>=0.1.1",
    "seaborn>=0.13.2",
    "wordcloud>=1.9.4",
    "llvmlite>=0.44.0",
    "bertopic>=0.16.4",
    "numpy==2.1.3",
    "transformers>=4.49.0",
    "underthesea>=6.8.4",
    "vncorenlp>=1.0.3",
    "protobuf==5.29.3",
    "tiktoken>=0.9.0",
    "sentencepiece>=0.2.0",
    "tensorflow>=2.19.0",
    "keras>=3.9.0",
    "instructor>=1.7.8",
    "python-dotenv>=1.1.0",
    "loguru>=0.7.3",
]
